+++ src/test/groovy/org/osiam/resources/scim/GroupSpec.groovy
@@ -32,16 +32,11 @@ class GroupSpec extends Specification {
     def 'should be able to generate a group'() {
         given:
 
-        User user = new User.Builder('userName').setId('some ID').build()
+        MemberRef memberRef = new MemberRef.Builder().setValue('id').build()
 
-        MemberRef memberRef = new MemberRef.Builder(user).build()
-
-        Meta meta = new Meta.Builder().build()
-
-        Group.Builder builder = new Group.Builder('display')
+        Group.Builder builder = new Group.Builder()
+                .setDisplayName('display')
                 .setExternalId('externalId')
-                .setId('id')
-                .setMeta(meta)
                 .setMembers([memberRef] as Set)
 
         when:
@@ -52,29 +47,41 @@ class GroupSpec extends Specification {
         group.displayName == 'display'
         group.externalId == 'externalId'
         group.schemas.first() == Group.SCHEMA
-        group.id == 'id'
     }
 
     def 'should be able to add member to group'() {
         when:
-        def group = new Group.Builder('display')
+        def group = new Group.Builder()
+                .setDisplayName('display')
                 .addMember(new MemberRef.Builder().build()).build()
 
         then:
         group.members.size() == 1
     }
 
+    def 'pre-set display name with constructor' () {
+        given:
+        def displayName = 'display name'
+
+        when:
+        def group = new Group.Builder(displayName)
+                .build()
+
+        then:
+        group.displayName == displayName
+    }
+
     def 'should be able to clone a group'() {
         given:
-        def group = new Group.Builder('display').
-                setId('id').build()
+        def group = new Group.Builder()
+                .setDisplayName('display')
+                .build()
         when:
         def result = new Group.Builder(group).build()
 
         then:
         group.displayName == result.displayName
         group.members == result.members
-        group.id == result.id
     }
 
     def 'using the copy-of builder with null as parameter raises exception'() {
@@ -85,9 +92,11 @@ class GroupSpec extends Specification {
         thrown(SCIMDataValidationException)
     }
 
-    def 'the copied group should have the given displayname'() {
+    def 'the copied group should have the given display name'() {
         given:
-        Group oldGroup = new Group.Builder("oldDisplayName").setExternalId("externalId").build()
+        Group oldGroup = new Group.Builder()
+                .setDisplayName('oldDisplayName')
+                .setExternalId('externalId').build()
         String newDisplayName = 'newDisplayName'
         Group newGroup
 
@@ -101,14 +110,11 @@ class GroupSpec extends Specification {
 
     def 'group can be serialized and deserialized'() {
         given:
-        User user = new User.Builder('userName').setId('some ID').build()
-        MemberRef memberRef = new MemberRef.Builder(user).build()
-        Meta meta = new Meta.Builder().build()
+        MemberRef memberRef = new MemberRef.Builder().setValue('id').build()
 
-        Group.Builder builder = new Group.Builder('display')
+        Group.Builder builder = new Group.Builder()
+                .setDisplayName('display')
                 .setExternalId('externalId')
-                .setId('id')
-                .setMeta(meta)
                 .setMembers([memberRef] as Set)
 
         Group group = builder.build()
@@ -119,7 +125,6 @@ class GroupSpec extends Specification {
         then:
         group == newGroup
         group.getExternalId() == newGroup.getExternalId()
-        group.getMeta() == newGroup.getMeta()
         group.getMembers().getAt(0) == newGroup.getMembers().getAt(0)
     }
 }