+++ src/test/groovy/org/osiam/resources/scim/UserJsonSpec.groovy
@@ -40,11 +40,8 @@ class UserJsonSpec extends Specification {
 
     def 'A User is correctly serialized'() {
         given:
-        User user = new User.Builder('bjensen')
-                .setId('a4bbe688-4b1e-4e4e-80e7-e5ba5c4d6db4')
-                .setMeta(new Meta.Builder()
-                        .setLocation('https://example.com/v1/Users/2819c223...')
-                        .setResourceType('User').build())
+        User user = new User.Builder()
+                .setUserName('bjensen')
                 .setExternalId('bjensen')
                 .setName(new Name.Builder()
                         .setFormatted('Ms. Barbara J Jensen III')
@@ -87,9 +84,6 @@ class UserJsonSpec extends Specification {
         JsonPath json = from(mapper.writeValueAsString(user))
 
         then:
-        json.getString('id') == 'a4bbe688-4b1e-4e4e-80e7-e5ba5c4d6db4'
-        json.getString('meta.location') == 'https://example.com/v1/Users/2819c223...'
-        json.getString('meta.resourceType') == 'User'
         json.getList('schemas').containsAll(['urn:ietf:params:scim:schemas:core:2.0:User',
                                              'urn:scim:schemas:extension:test:1.0:User'])
         json.getString('externalId') == 'bjensen'